JWT을 이용한 로그인 기능 추가. 전화번호부 목록 추가 API 추가

+++ controllers/user.js
@@ -1,6 +1,41 @@
+const jwt = require('jsonwebtoken');
 const User = require('../models/user');
 
 module.exports = {
+    loginUser: async (req, res, next) => {
+        try {
+            if(!req.body.id.trim() || !req.body.password.trim()) {
+                return res.status(400).json({error: 'Unexpected JSON input'});
+            }
+
+            const id = req.body.id;
+            const pw = req.body.password;
+            const user = await User.findOne({ID: id});
+            if(user) {
+                if(pw == user.Password) {
+                    const token = jwt.sign({
+                        id: user._id
+                    },
+                    'BaeMyunghoCadaWord',{
+                        expiresIn: '1h'
+                    });
+
+                    return res.status(200).json({token: token});
+                }
+                else {
+                    return res.status(401).json({error: 'Password is incorrect'});
+                }
+            }
+            else {
+                return res.status(404).json({error: 'No user data to get key'});
+            }
+        }
+        catch(err) {
+            console.error(err);
+            return res.status(500).json({error: 'Internal server error'});           
+        }
+    },
+
     createUser: async (req, res, next) => {
         try {
             if(!req.body.id.trim() || !req.body.password.trim() || !req.body.nickname) {
@@ -55,6 +90,10 @@ module.exports = {
 
     getUserById: async (req, res, next) => {
         try {
+            if(!req.isAuth) {
+                return res.status(401).json({error: 'Unauthenticated'});
+            }
+
             const id = req.params.id;
             const user = await User.findOne({ID: id});
             if(user) {
@@ -72,6 +111,10 @@ module.exports = {
 
     getUserByNickname: async (req, res, next) => {
         try {
+            if(!req.isAuth) {
+                return res.status(401).json({error: 'Unauthenticated'});
+            }
+
             const nickname = req.params.nickname;
             const user = await User.findOne({Nickname: nickname});
             if(user) {
@@ -93,6 +136,10 @@ module.exports = {
 
     getUsers: async (req, res, next) => {
         try {
+            if(!req.isAuth) {
+                return res.status(401).json({error: 'Unauthenticated'});
+            }
+            
             const users = await User.find();
             if(users) {
                 return res.status(200).json(users);
@@ -109,6 +156,10 @@ module.exports = {
 
     updateUserById: async (req, res, next) => {
         try {
+            if(!req.isAuth) {
+                return res.status(401).json({error: 'Unauthenticated'});
+            }
+
             const id = req.params.id;
             const user = await User.findOne({ID: id});
             if(user) {
@@ -146,6 +197,10 @@ module.exports = {
 
     updateUserByNickname: async (req, res, next) => {
         try {
+            if(!req.isAuth) {
+                return res.status(401).json({error: 'Unauthenticated'});
+            }
+
             const nickname = req.params.nickname;
             const user = await User.findOne({Nickname: nickname});
             if(user) {
@@ -157,7 +212,7 @@ module.exports = {
                 if(searchedUser.length) {
                     return res.status(409).json({error: 'ID or Nickname is already exist'});
                 }
-                
+
                 user.Password = req.body.password.trim();
                 user.Nickname = req.body.nickname;
                 user.Comment = req.body.comment;
@@ -183,6 +238,10 @@ module.exports = {
 
     deleteUserById: async (req, res, next) => {
         try {
+            if(!req.isAuth) {
+                return res.status(401).json({error: 'Unauthenticated'});
+            }
+
             const id = req.params.id;
             const deletedUser = await User.findOneAndDelete({ID: id});
             if(deletedUser) {
@@ -200,6 +259,10 @@ module.exports = {
 
     deleteUserByNickname: async (req, res, next) => {
         try {
+            if(!req.isAuth) {
+                return res.status(401).json({error: 'Unauthenticated'});
+            }
+
             const nickname = req.params.nickname;
             const deletedUser = await User.findOneAndDelete({Nickname: nickname});
             if(deletedUser) {
